Filing Wizard
A Python program called the Filing Wizard is intended to assist you with automatically categorizing files in your "Downloads" folder into separate subfolders according to their type of file. This may minimize congestion and make it easier to locate things in your Downloads folder by maintaining it organized and tidy.